added SearchField component & its styles to the components folder,also add react-icon library to render search icon inside the input field, also update the App component to fetch contactList data from third party server
