--- src/App.js
@@ -1,18 +1,0 @@
-import React from 'react'
-import './App.css';
-
-class App extends React.Component{
-  render(){
-    return (
-      <div className="App">
-        <header className="App-header">
-          <h1>Contacts</h1>
-        </header>
-      </div>
-    );
-
-  }
-
-}
-
-export default App;
