+++ src/Components/searchField/searchField.css
@@ -1,0 +1,22 @@
+/* searchField container */
+.searchField-container {
+  width: 100%;
+  position: relative;
+}
+
+/* searchfield input */
+.searchField-container input {
+  width: 100%;
+  padding: 0.6rem 0.5rem;
+  /* border: none; */
+  padding-left: 1.2rem;
+}
+
+/* search-icon */
+.search-icon {
+  position: absolute;
+  top: 0.7rem;
+  left: 1rem;
+  font-size: 1rem;
+  color: rgb(192, 190, 190);
+}
